Editar curso finalizado

+++ objetos/c_cursos/c_cursos.h
@@ -25,7 +25,7 @@ class Ccurso : public Usuario{
 	bool crearCurso();
 	bool editarCurso(int ncur);
 	bool eliminarCurso();
-	int comprobarExistencia(string c);//devuelve la posicion de la linea donde se encuentra el elemento existente y -1 si no existe
+	size_t comprobarExistencia(string c);//devuelve la posicion de la linea donde se encuentra el elemento existente y -1 si no existe
 	string estudiosRequeridos();
 	string fecha(int mod);
